Consolidating the Tests and removing the external Defaults flag.
