+++ tests/ga/test_extension.py
@@ -16,14 +16,11 @@
 #
 
 import os.path
-import sys
 
-from azurelinuxagent.common.cgroups.cgutils import CGroupsException
-from azurelinuxagent.common.exception import ExtensionHandlerConfigurationError
 from azurelinuxagent.common.protocol.restapi import Extension
 from azurelinuxagent.common.protocol.wire import WireProtocol, InVMArtifactsProfile
 from azurelinuxagent.ga.exthandlers import *
-from azurelinuxagent.ga.utils.exthandler_utils import HandlerConfiguration, CGroupsLimits
+from azurelinuxagent.ga.utils.exthandler_utils import CGroupsLimits
 from tests.common.test_cgroups import i_am_root
 from tests.protocol.mockwiredata import *
 
@@ -1855,538 +1852,6 @@ class TestInVMArtifactsProfile(AgentTestCase):
         profile_json = '{ "onHold": "TRUE" }'
         profile = InVMArtifactsProfile(profile_json)
         self.assertTrue(profile.is_on_hold(), "Failed to parse '{0}'".format(profile_json))
-
-
-class TestHandlerConfiguration(ExtensionTestCase):
-
-    def test_handler_configuration_complete_configuration(self, *args):
-        name = "ExampleHandlerLinux"
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 2,
-                "limit_percentage": 25
-              },
-              {
-                "cores": 8,
-                "limit_percentage": 20
-              },
-              {
-                "cores": -1,
-                "limit_percentage": 15
-              }
-            ],
-            "memory": {
-              "max_limit_percentage": 20,
-              "max_limit_MBs": 1000,
-              "memory_pressure_warning": "low",
-              "memory_oom_kill": "enabled"
-            }
-          }
-        },
-        "windows": {}
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name)
-        resource_config = handler_config.get_resource_configurations()
-
-        self.assertEqual(handler_config.get_name(), "ExampleHandlerLinux")
-        self.assertEqual(handler_config.get_version(), 1.0)
-
-        cpu_limits = resource_config.get_cpu_limits_for_extension().cpu_limits
-        self.assertEqual(3, len(cpu_limits))
-
-        memory_limits = resource_config.get_memory_limits_for_extension()
-        self.assertEqual(memory_limits.max_limit_percentage, 20)
-        self.assertEqual(memory_limits.max_limit_MBs, 1000)
-        self.assertEqual(memory_limits.memory_pressure_warning, "low")
-        self.assertEqual(memory_limits.memory_oom_kill, "enabled")
-
-    def test_handler_configuration_only_cpu_configuration(self, *args):
-        name = "ExampleHandlerLinux"
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 2,
-                "limit_percentage": 25
-              },
-              {
-                "cores": 8,
-                "limit_percentage": 20
-              },
-              {
-                "cores": -1,
-                "limit_percentage": 15
-              }
-            ]
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        resource_config = handler_config.get_resource_configurations()
-
-        self.assertEqual(resource_config.get_memory_limits_for_extension(), None)
-
-        cpu_limits = resource_config.get_cpu_limits_for_extension().cpu_limits
-        self.assertEqual(3, len(cpu_limits))
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": -1,
-                "limit_percentage": 15
-              }
-            ]
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        resource_config = handler_config.get_resource_configurations()
-
-        self.assertEqual(resource_config.get_memory_limits_for_extension(), None)
-
-        cpu_limits = resource_config.get_cpu_limits_for_extension().cpu_limits
-        self.assertEqual(1, len(cpu_limits))
-        self.assertEqual(-1, cpu_limits[0].cores)
-        self.assertEqual(15, cpu_limits[0].limit_percentage)
-
-    def test_handler_configuration_incorrect_cpu_configuration(self, *args):
-        name = "ExampleHandlerLinux"
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 2,
-                "limit_percentage": 25
-              },
-              {
-                "cores": 8,
-                "limit_percentage": 20
-              }
-            ]
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 2,
-                "limit_percentage": 25
-              },
-              {
-                "cores": 8
-              }
-            ]
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 0,
-                "limit_percentage": 25
-              },
-              {
-                "cores": -1,
-                "limit_percentage": 25
-              }
-            ]
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 2,
-                "limit_percentage": 251
-              },
-              {
-                "cores": 8,
-                "limit_percentage": 20
-              },
-              {
-                "cores": -1,
-                "limit_percentage": 15
-              }
-            ]
-          }
-        }
-      }
-    }
-            '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 2,
-                "limit_percentage": 0.25
-              },
-              {
-                "cores": 8,
-                "limit_percentage": 20
-              },
-              {
-                "cores": -1,
-                "limit_percentage": 15
-              }
-            ]
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "cpu": [
-              {
-                "cores": 2,
-                "limit_percentage": 25
-              },
-              {
-                "cores": 8,
-                "limit_percentage": 20
-              },
-              {
-                "cores": -1,
-                "limit_percentage": "15"
-              }
-            ]
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-    def test_handler_configuration_only_memory_configuration(self, *args):
-        name = "ExampleHandlerLinux"
-        data = '''{
-          "name": "ExampleHandlerLinux",
-          "version": 1.0,
-          "handlerConfiguration": {
-            "linux": {
-              "resources": {
-                "memory": {
-                  "max_limit_percentage": 20,
-                  "max_limit_MBs": 1000,
-                  "memory_pressure_warning": "low",
-                  "memory_oom_kill": "enabled"
-                }
-              }
-            }
-          }
-        }
-        '''
-        resource_config = HandlerConfiguration(json.loads(data), name=name)
-        resource_config = resource_config.get_resource_configurations()
-
-        self.assertEqual(resource_config.get_cpu_limits_for_extension(), None)
-
-        memory_limits = resource_config.get_memory_limits_for_extension()
-        self.assertEqual(memory_limits.max_limit_percentage, 20)
-        self.assertEqual(memory_limits.max_limit_MBs, 1000)
-        self.assertEqual(memory_limits.memory_pressure_warning, "low")
-        self.assertEqual(memory_limits.memory_oom_kill, "enabled")
-
-        data = '''{
-          "name": "ExampleHandlerLinux",
-          "version": 1.0,
-          "handlerConfiguration": {
-            "linux": {
-              "resources": {
-                "memory": {
-                  "max_limit_percentage": 20,
-                  "max_limit_MBs": 1000,
-                  "memory_pressure_warning": "low"
-                }
-              }
-            }
-          }
-        }
-                '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        resource_config = handler_config.get_resource_configurations()
-
-        memory_limits = resource_config.get_memory_limits_for_extension()
-        self.assertEqual(memory_limits.max_limit_percentage, 20)
-        self.assertEqual(memory_limits.max_limit_MBs, 1000)
-        self.assertEqual(memory_limits.memory_pressure_warning, "low")
-        self.assertEqual(memory_limits.memory_oom_kill, None)  # Default is set by CGroupsLimits class.
-
-        data = '''{
-          "name": "ExampleHandlerLinux",
-          "version": 1.0,
-          "handlerConfiguration": {
-            "linux": {
-              "resources": {
-                "memory": {
-                  "max_limit_percentage": 20,
-                  "max_limit_MBs": 1000
-                }
-              }
-            }
-          }
-        }
-        '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        resource_config = handler_config.get_resource_configurations()
-
-        memory_limits = resource_config.get_memory_limits_for_extension()
-        self.assertEqual(memory_limits.max_limit_percentage, 20)
-        self.assertEqual(memory_limits.max_limit_MBs, 1000)
-        self.assertEqual(memory_limits.memory_pressure_warning, None)
-        self.assertEqual(memory_limits.memory_oom_kill, None)  # Default is set by CGroupsLimits class.
-
-    def test_handler_configuration_incorrect_memory_configuration(self, *args):
-        name = "ExampleHandlerLinux"
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "memory": {
-              "max_limit_percentage": 20,
-              "max_limit_MBs": 1000,
-              "memory_pressure_warning": "low",
-              "memory_oom_kill": "enabledd"
-            }
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "memory": {
-              "max_limit_percentage": 200,
-              "max_limit_MBs": 1000
-            }
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "memory": {
-              "max_limit_percentage": "20",
-              "max_limit_MBs": 1000
-            }
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-      "name": "ExampleHandlerLinux",
-      "version": 1.0,
-      "handlerConfiguration": {
-        "linux": {
-          "resources": {
-            "memory": {
-              "max_limit_percentage": 20,
-              "max_limit_MBs": "1000"
-            }
-          }
-        }
-      }
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-    def test_handler_configuration_incorrect_configuration(self, *args):
-        name = "ExampleHandlerLinux"
-        data = '''{
-          "name": "ExampleHandlerLinux",
-          "version": 1.0,
-          "handlerConfiguration": {
-            "linux": {
-              "resources": {
-              }
-            }
-          }
-        }
-        '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-                  "name": "ExampleHandlerLinux",
-                  "version": 1.0,
-                  "handlerConfiguration": {}
-                }
-                '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-        data = '''{
-    "name": "ExampleHandlerLinux",
-    "version": 1.0
-    }
-    '''
-        handler_config = HandlerConfiguration(json.loads(data), name=name)
-        self.assertIsNotNone(handler_config)
-        self.assertIsNone(handler_config.get_resource_configurations())
-        self.assertIsNone(handler_config.get_resource_limits())
-
-    @patch("azurelinuxagent.ga.exthandlers.HandlerConfiguration.send_handler_configuration_event")
-    def test_load_handler_configuration(self, patch_send_handler_configuration_event, *args):
-        handler_name = "Not.A.Real.Extension"
-        handler_version = "1.2.3"
-
-        ext_handler = ExtHandler(handler_name)
-        ext_handler.properties.version = handler_version
-        ext_handler_i = ExtHandlerInstance(ext_handler, "dummy protocol")
-
-        with patch(
-                "azurelinuxagent.ga.exthandlers.ExtHandlerInstance.get_handler_configuration_file") as \
-                patch_get_handler_configuration_file:
-            patch_get_handler_configuration_file.return_value = load_data_path("ext/SampleHandlerConfiguration.json")
-            handler_config = ext_handler_i.load_handler_configuration()
-            self.assertNotEqual(handler_config, None)
-            self.assertEqual(patch_send_handler_configuration_event.call_count, 0)
-
-        with patch(
-                "azurelinuxagent.ga.exthandlers.ExtHandlerInstance.get_handler_configuration_file") as \
-                patch_get_handler_configuration_file:
-            patch_get_handler_configuration_file.return_value = load_data_path("ext/NotPresent.json")
-            handler_config = ext_handler_i.load_handler_configuration()
-            self.assertEqual(handler_config, None)
-            self.assertEqual(patch_send_handler_configuration_event.call_count, 0)
-
-        with patch(
-                "azurelinuxagent.ga.exthandlers.ExtHandlerInstance.get_handler_configuration_file") as \
-                patch_get_handler_configuration_file:
-            patch_get_handler_configuration_file.return_value = load_data_path(
-                "ext/SampleMalformedHandlerConfiguration.json")
-            handler_config = ext_handler_i.load_handler_configuration()
-            self.assertEqual(handler_config, None)
-            self.assertEqual(patch_send_handler_configuration_event.call_count, 1)
-
-        with patch(
-                "azurelinuxagent.ga.exthandlers.ExtHandlerInstance.get_handler_configuration_file") as \
-                patch_get_handler_configuration_file:
-            patch_get_handler_configuration_file.return_value = load_data_path(
-                "ext/SampleInvalidButCorrectJsonHandlerConfiguration.json")
-            handler_config = ext_handler_i.load_handler_configuration()
-            self.assertNotEqual(handler_config, None)
-            self.assertEqual(patch_send_handler_configuration_event.call_count, 3)
 
 
 if __name__ == '__main__':
